fix font size and position of scroll bar scalability issues

+++ src/main/Size.java
@@ -9,7 +9,7 @@ import javafx.geometry.Insets;
 
 public class Size {
 
-    public static double fontSize(int i) {
+    public static double fontSize(double i) {
         return Root.getUtilAndConstants().fontSize(i);
     }
